requirements.txt
streamlit==1.39.0
PyMySQL==1.1.1
pandas==2.2.2
python-docx==1.1.2
face-recognition==1.3.0
opencv-python==4.8.1.78
numpy==1.26.4
# dlib==19.24.6
# playsound
# python-dateutil==2.9.0.post0
cryptography
openpyxl

# pip install --upgrade setuptools
# pip install git+https://github.com/ageitgey/face_recognition_models
# pip install opencv-python
